requirements.txt
fastapi~=0.104.1
uvicorn~=0.24.0
python-dotenv~=0.19.0
elevenlabs~=0.2.24
python-multipart~=0.0.6
openai-whisper
numpy
sounddevice